refactor(auth): collapse org and project seams into one enforceAuth
