+++ web/src/pages/api/public/projects/[projectId]/index.ts
@@ -35,7 +35,6 @@ export default async function handler(
   // CHECK AUTH
   const authCheck = await verifyOrgAuth({
     req,
-    projectId,
     action: req.method === "PUT" ? "project:update" : "project:delete",
   });
   if (!authCheck.validKey) {
